Code Cleaning
Cleaning up some code.
Changed old Note and Measure classes to RawNote and RawMeasure to be
used by the sound generator to generate the sound.
Created new Note and Measure to be used with the generator to allow for
less confusing stuff, namely from the ending of notes, and allowing for
notes to be different lengths.
TODO:
-Fix up the player to support ending notes after the end of the measure
and not get caught up on waiting for a note to end
- Make measure "length" more concrete (i.e. saying that a measure is 4/4
with 4 subdivisions per beat.)

Didn't finish so lot's of things will be broken.